refactor(frontend): Move ArchiveHeader component directory
